Add files via upload
Cell.qml and main.qml contain the qml code to build an "app"
Cell.qml contains relevant dimensions and properties of a "cell" that's visible on main.qml with different colours. Clicking on said colours changes the colour of the text on main.qml. 
Clicking on the text in main.qml, changed the text to show proof of concept for changing data values required for hyperloop